Seperating out Charts
