+++ src/main/java/ledger/user_interface/ui_controllers/window/ExpenditureChartsController.java
@@ -1,28 +1,26 @@
 package ledger.user_interface.ui_controllers.window;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.chart.LineChart;
-import javafx.scene.chart.PieChart;
-import javafx.scene.chart.XYChart;
 import javafx.scene.control.Button;
 import javafx.scene.control.DatePicker;
 import javafx.scene.layout.GridPane;
 import ledger.controller.DbController;
 import ledger.controller.register.TaskWithReturn;
 import ledger.database.entity.Account;
-import ledger.database.entity.Tag;
 import ledger.database.entity.Transaction;
 import ledger.user_interface.ui_controllers.IUIController;
 import ledger.user_interface.ui_controllers.component.AccountDropdown;
+import ledger.user_interface.ui_controllers.component.charts.ExpenditureLineChart;
+import ledger.user_interface.ui_controllers.component.charts.ExpenditurePieChart;
 
 import java.net.URL;
-import java.text.*;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+import java.util.ResourceBundle;
 
 /**
  * Controls how the charts render with user given information.
@@ -36,9 +34,9 @@ public class ExpenditureChartsController extends GridPane implements Initializab
     @FXML
     private DatePicker toDateFilter;
     @FXML
-    private PieChart expendituresPieChart;
+    private ExpenditurePieChart expendituresPieChart;
     @FXML
-    private LineChart expendituresLineChart;
+    private ExpenditureLineChart expendituresLineChart;
     @FXML
     private Button filterEnterButton;
 
@@ -48,8 +46,8 @@ public class ExpenditureChartsController extends GridPane implements Initializab
     ExpenditureChartsController() {
         this.initController(pageLoc, this, "Error on expenditure chart page startup: ");
         getTransactions();
-        setupExpenditureHistoryChart();
-        setupInitialPieChart();
+        expendituresLineChart.updateData(this.allTransactions);
+        expendituresPieChart.updateData(this.allTransactions);
     }
 
     /**
@@ -105,98 +103,6 @@ public class ExpenditureChartsController extends GridPane implements Initializab
     }
 
     /**
-     * Builds the line chart to show trends in amount spent over the last six months
-     */
-    private void setupExpenditureHistoryChart() {
-        Calendar cal = Calendar.getInstance();
-        Date today = cal.getTime();
-        cal.add(Calendar.MONTH, -6);
-        Date sixMonthsAgo = cal.getTime();
-        List<Transaction> filteredTransactions = new ArrayList<>();
-        for (Transaction t : this.allTransactions) {
-            if ((t.getDate().before(today) || t.getDate().equals(today))
-                    && (t.getDate().after(sixMonthsAgo) || t.getDate().equals(sixMonthsAgo))) {
-                filteredTransactions.add(t);
-            }
-        }
-
-        this.expendituresLineChart.getXAxis().setLabel("Month");
-        this.expendituresLineChart.getYAxis().setLabel("Net Expenditure");
-        this.expendituresLineChart.getXAxis().setAutoRanging(true);
-        this.expendituresLineChart.getYAxis().setAutoRanging(true);
-
-        Map<String, Long> monthToAmountSpent = new HashMap<>();
-        Map<String, Integer> monthToYear = new HashMap<>();
-        for (Transaction t : filteredTransactions) {
-            cal.setTime(t.getDate());
-            String month = new DateFormatSymbols().getMonths()[cal.get(Calendar.MONTH)];
-            Integer year = cal.get(Calendar.YEAR);
-            monthToYear.put(month, year);
-            addToMapForLineChart(monthToAmountSpent, month, t.getAmount());
-        }
-
-        Set<String> months = monthToAmountSpent.keySet();
-        List<String> preorderedMonths = new ArrayList<>();
-        preorderedMonths.addAll(months);
-
-        orderMonthsAndYears(monthToYear, preorderedMonths);
-
-        XYChart.Series series = new XYChart.Series();
-        for (String m : preorderedMonths) {
-            series.getData().add(new XYChart.Data(m, monthToAmountSpent.get(m) / 100));
-        }
-        series.setName("Change in Account Balance");
-        this.expendituresLineChart.getData().add(series);
-        this.expendituresLineChart.setVisible(true);
-    }
-
-    /**
-     * Takes care of ordering the months chronologically and also handles the switch
-     * from December to January in a new year
-     *
-     * @param monthToYear      HashMap that keeps references from each month to their respective year
-     * @param preorderedMonths Unordered list of months
-     */
-    private void orderMonthsAndYears(Map<String, Integer> monthToYear, List<String> preorderedMonths) {
-        preorderedMonths.sort((String o1, String o2) -> {
-            SimpleDateFormat s = new SimpleDateFormat("MMM");
-            try {
-                Date s1 = s.parse(o1);
-                Date s2 = s.parse(o2);
-                return s1.compareTo(s2);
-            } catch (ParseException e) {
-                e.printStackTrace();
-            }
-            return -1;
-        });
-        // takes care of transitioning into a new year
-        List<String> monthsInNextYear = new ArrayList<>();
-        Integer lowestYear = 0;
-        for (int j = 0; j < preorderedMonths.size(); j++) {
-            if (j == 0) {
-                lowestYear = monthToYear.get(preorderedMonths.get(j));
-            } else if (monthToYear.get(preorderedMonths.get(j)) < lowestYear) {
-                lowestYear = monthToYear.get(preorderedMonths.get(j));
-            }
-        }
-        for (String preorderedMonth : preorderedMonths) {
-            if (monthToYear.get(preorderedMonth) > lowestYear) {
-                monthsInNextYear.add(preorderedMonth);
-            }
-        }
-        // takes the months in the next year out of the beginning of the list and tacks them on the end
-        preorderedMonths.removeAll(monthsInNextYear);
-        preorderedMonths.addAll(monthsInNextYear);
-    }
-
-    /**
-     * Sets up PieChart seen on initialization of expenditures screen. Data is loaded from all accounts.
-     */
-    private void setupInitialPieChart() {
-        createPieChart(this.allTransactions);
-    }
-
-    /**
      * Creates the stacked bar chart based on transactions in a specified account
      *
      * @param account the account to filter the transactions by
@@ -209,7 +115,7 @@ public class ExpenditureChartsController extends GridPane implements Initializab
             }
         }
 
-        createPieChart(filteredTransactions);
+        expendituresPieChart.updateData(filteredTransactions);
     }
 
     /**
@@ -230,7 +136,7 @@ public class ExpenditureChartsController extends GridPane implements Initializab
             }
         }
 
-        createPieChart(filteredTransactions);
+        expendituresPieChart.updateData(filteredTransactions);
     }
 
     /**
@@ -258,87 +164,6 @@ public class ExpenditureChartsController extends GridPane implements Initializab
             }
         }
 
-        createPieChart(filteredTransactions);
-    }
-
-    /**
-     * Creates a map of tag to amount spent on that tag, and then uses the map to create a Pie chart
-     *
-     * @param filteredTransactions transactions fitting the criteria of the filter
-     */
-    private void createPieChart(List<Transaction> filteredTransactions) {
-        Map<String, Long> tagNameToAmountSpent = new HashMap<>();
-        for (Transaction t : filteredTransactions) {
-            if (t.getTags().isEmpty()) {
-                addToMapForPieChart(tagNameToAmountSpent, "Uncategorized", t.getAmount());
-            } else {
-                for (Tag tag : t.getTags()) {
-                    addToMapForPieChart(tagNameToAmountSpent, tag.getName(), t.getAmount());
-                }
-            }
-        }
-        List<PieChart.Data> dataList = new ArrayList<>();
-        for (String tag : tagNameToAmountSpent.keySet()) {
-            // use absolute value here so it's not negative
-            double amountSpent = Math.abs(tagNameToAmountSpent.get(tag)) / 100;
-            NumberFormat formatter = new DecimalFormat("#0.00");
-            dataList.add(new PieChart.Data(tag + " - " + "(" + formatter.format(amountSpent) + ")", amountSpent));
-        }
-        if (dataList.isEmpty()) {
-            this.setupErrorPopup("There's no data to be displayed!", new Exception());
-            return;
-        }
-        ObservableList<PieChart.Data> pieChartData = FXCollections.observableArrayList(dataList);
-        this.expendituresPieChart.setData(pieChartData);
-        this.expendituresPieChart.setTitle("Expenditures by Category");
-        this.expendituresPieChart.setVisible(true);
-
-    }
-
-    /**
-     * Used to determine what filtered data should show on expenditure pie chart
-     * Only adds up amount spent (minus dollars) on items - doesn't consider deposits
-     *
-     * @param map   map in which the data is organized
-     * @param key   map key to check value
-     * @param value value to add to existing value or empty map
-     */
-    private void addToMapForPieChart(Map map, String key, Long value) {
-        if (value <= 0) {
-            if (key.equals("")) {
-                key = "Uncategorized";
-            }
-            populateMap(map, key, value);
-        }
-    }
-
-    /**
-     * Used to determine what filtered data should show on expenditure line chart
-     * Adds up amount spent and considers deposits for a net account balance change
-     *
-     * @param map   map in which the data is organized
-     * @param key   map key to check value
-     * @param value value to add to existing value or empty map
-     */
-    private void addToMapForLineChart(Map map, String key, Long value) {
-        populateMap(map, key, value);
-    }
-
-    /**
-     * Populates the map passed in with given key and value
-     *
-     * @param map   map in which the data is organized
-     * @param key   map key to check value
-     * @param value value to add to existing value or empty map
-     */
-    private void populateMap(Map map, String key, Long value) {
-        if (!map.keySet().contains(key)) {
-            map.put(key, value);
-        } else {
-            Long existingAmount = (Long) map.get(key);
-            existingAmount += value;
-
-            map.put(key, existingAmount);
-        }
+        expendituresPieChart.updateData(filteredTransactions);
     }
 }
