+++ src/main/resources/fxml_files/ExpenditureCharts.fxml
@@ -16,6 +16,8 @@
 <?import javafx.scene.text.Text?>
 <?import ledger.user_interface.ui_controllers.component.AccountDropdown?>
 
+<?import ledger.user_interface.ui_controllers.component.charts.ExpenditureLineChart?>
+<?import ledger.user_interface.ui_controllers.component.charts.ExpenditurePieChart?>
 <fx:root fx:id="mainBackground" maxHeight="800.0" maxWidth="1280.0" minHeight="800.0" minWidth="1280.0" prefHeight="800.0" prefWidth="1280.0" style="-fx-background-color: #ffffff;" type="GridPane" xmlns="http://javafx.com/javafx/8.0.112" xmlns:fx="http://javafx.com/fxml/1">
     <columnConstraints>
         <ColumnConstraints hgrow="SOMETIMES" maxWidth="0.0" minWidth="178.0" percentWidth="50.0" prefWidth="178.0" />
@@ -83,7 +85,7 @@
                </GridPane.margin></Button>
          </children>
       </GridPane>
-      <LineChart fx:id="expendituresLineChart" prefHeight="156.0" prefWidth="135.0" title="Expenditure History" GridPane.rowIndex="2">
+      <ExpenditureLineChart fx:id="expendituresLineChart" prefHeight="156.0" prefWidth="135.0" title="Expenditure History" GridPane.rowIndex="2">
         <xAxis>
           <CategoryAxis side="BOTTOM" />
         </xAxis>
@@ -96,12 +98,12 @@
          <GridPane.margin>
             <Insets top="10.0" />
          </GridPane.margin>
-      </LineChart>
-      <PieChart fx:id="expendituresPieChart" animated="false" GridPane.columnIndex="1" GridPane.rowIndex="2">
+      </ExpenditureLineChart>
+      <ExpenditurePieChart fx:id="expendituresPieChart" animated="false" GridPane.columnIndex="1" GridPane.rowIndex="2">
          <GridPane.margin>
             <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
          </GridPane.margin>
-      </PieChart>
+      </ExpenditurePieChart>
     </children>
     <stylesheets>
         <URL value="@../css/colorStyle.css" />
